+++ src/surrealist/ql/statements/remove.py
@@ -9,14 +9,13 @@ REMOVE [
     NAMESPACE [ IF EXISTS ] @name
     | DATABASE [ IF EXISTS] @name
     | USER [ IF EXISTS ] @name ON [ ROOT | NAMESPACE | DATABASE ]
-    | TOKEN [ IF EXISTS ] @name ON [ NAMESPACE | DATABASE ]
+    | ACCESS [ IF EXISTS ] @name ON [ NAMESPACE | DATABASE ]
     | EVENT [ IF EXISTS ] @name ON [ TABLE ] @table
     | FIELD [ IF EXISTS ] @name ON [ TABLE ] @table
     | INDEX [ IF EXISTS ] @name ON [ TABLE ] @table
     | ANALYZER [ IF EXISTS ] @name
     | FUNCTION [ IF EXISTS ] fn::@name
     | PARAM [ IF EXISTS ] $@name
-    | SCOPE [ IF EXISTS ] @name
     | TABLE [ IF EXISTS ] @name
 ]
 """
@@ -30,7 +29,7 @@ class Remove(Statement):
 
     Examples: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/database.py
     """
-    _variants = ("TABLE", "EVENT", "FIELD", "INDEX", "PARAM", "USER", "ANALYZER", "SCOPE", "TOKEN")
+    _variants = ("TABLE", "EVENT", "FIELD", "INDEX", "PARAM", "USER", "ANALYZER", "SCOPE", "TOKEN", "ACCESS")
 
     def __init__(self, connection: Connection, table_name: str, type_: str = "TABLE", name: Optional[str] = None):
         if type_ not in Remove._variants:
@@ -61,9 +60,9 @@ class Remove(Statement):
 
     def _clean_str(self):
         add = "" if not self._on_exists else " IF EXISTS"
-        if self._type in ("TABLE", "PARAM", "USER", "ANALYZER", "SCOPE", "TOKEN"):
+        if self._type in ("TABLE", "PARAM", "USER", "ANALYZER", "SCOPE", "TOKEN", "ACCESS"):
             what = f"{self._type}{add} {self._name}"
-            if self._type in ("USER", "TOKEN"):
+            if self._type in ("USER", "TOKEN", "ACCESS"):
                 what = f"{what} ON DATABASE"
             if self._type == "PARAM":
                 what = f"{self._type}{add} ${self._name}"